+++ src/carousel.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import IconChevronLeftRegular from './generated/mistica-icons/icon-chevron-left-regular';
 import IconChevronRightRegular from './generated/mistica-icons/icon-chevron-right-regular';
-import {useScreenSize, useTheme} from './hooks';
+import {useIsInViewport, useScreenSize, useTheme} from './hooks';
 import Inline from './inline';
 import {createUseStyles} from './jss';
 import Stack from './stack';
@@ -12,9 +12,17 @@ import {useIsInverseVariant, ThemeVariant} from './theme-variant-context';
 import {applyAlpha} from './utils/color';
 import {DisableBorderRadiusProvider} from './image';
 import {getPrefixedDataAttributes, listenResize} from './utils/dom';
+import {isAndroid} from './utils/platform';
+import {useDocumentVisibility} from './utils/document-visibility';
 
 import type {DataAttributes} from './utils/types';
 import type {Theme} from './theme';
+
+const useShouldAutoplay = (autoplay: boolean, ref: React.RefObject<HTMLElement>): boolean => {
+    const isDocumentVisible = useDocumentVisibility();
+    const isInViewport = useIsInViewport(ref, false);
+    return isInViewport && isDocumentVisible && !!autoplay;
+};
 
 const useBulletsStyles = createUseStyles((theme) => ({
     bullet: {
@@ -138,6 +146,7 @@ const useStyles = createUseStyles((theme) => ({
         },
     },
     item: {
+        scrollSnapStop: isAndroid(theme.platformOverrides) ? 'always' : 'normal',
         scrollSnapAlign: 'start',
         flexShrink: 0,
         width: ({itemsPerPageConfig, gap}) =>
@@ -282,7 +291,7 @@ const BaseCarousel: React.FC<BaseCarouselProps> = ({
     itemsPerPage,
     itemsToScroll,
     mobilePageOffset = 16,
-    gap = 8,
+    gap: gapProp,
     free,
     centered,
     autoplay,
@@ -293,6 +302,7 @@ const BaseCarousel: React.FC<BaseCarouselProps> = ({
     const itemsPerPageConfig = normalizeItemsPerPage(itemsPerPage);
     const mobilePageOffsetConfig = normalizeMobilePageOffset(mobilePageOffset);
     const {isDesktopOrBigger} = useScreenSize();
+    const gap: number = gapProp ?? (isDesktopOrBigger ? 16 : 8);
     const sideMargin = useResonsiveLayoutMargin();
     const classes = useStyles({itemsPerPageConfig, mobilePageOffsetConfig, free, gap, centered, sideMargin});
     const carouselRef = React.useRef<HTMLDivElement>(null);
@@ -327,13 +337,26 @@ const BaseCarousel: React.FC<BaseCarouselProps> = ({
             const calcItemScrollPositions = () => {
                 const maxScroll = carouselEl.scrollWidth - carouselEl.clientWidth;
 
+                const getItemScrollMargin = (itemIndex: number) => {
+                    if (centered) {
+                        return 0;
+                    }
+                    if (itemIndex === 0) {
+                        return 0;
+                    }
+                    if (isDesktopOrBigger) {
+                        return -gap;
+                    }
+                    return mobilePageOffsetConfig.prev;
+                };
+
                 setItemScrollPositions(
                     Array.from(carouselEl.querySelectorAll('[data-item]')).map((itemEl, idx) => {
                         if (idx === items.length - 1) {
                             return maxScroll;
                         }
                         const offsetLeft = (itemEl as HTMLElement).offsetLeft;
-                        const scrollMargin = Number(getComputedStyle(itemEl).scrollMargin.replace('px', ''));
+                        const scrollMargin = getItemScrollMargin(idx);
                         const scrollPosition =
                             centered && !isDesktopOrBigger ? offsetLeft - itemEl.clientWidth / 2 : offsetLeft;
                         return Math.min(scrollPosition - scrollMargin - carouselEl.offsetLeft, maxScroll);
@@ -400,8 +423,10 @@ const BaseCarousel: React.FC<BaseCarouselProps> = ({
         }
     }, [scrollPositions]);
 
+    const shouldAutoplay = useShouldAutoplay(!!autoplay, carouselRef);
+
     React.useEffect(() => {
-        if (autoplay) {
+        if (shouldAutoplay && autoplay) {
             const time = typeof autoplay === 'boolean' ? DEFAULT_AUTOPLAY_TIME : autoplay.time;
             const loop = typeof autoplay === 'object' && autoplay.loop;
             const interval = setInterval(() => {
@@ -413,7 +438,7 @@ const BaseCarousel: React.FC<BaseCarouselProps> = ({
             }, time);
             return () => clearInterval(interval);
         }
-    }, [autoplay, goNext, scrollRight]);
+    }, [autoplay, goNext, scrollRight, shouldAutoplay]);
 
     const currentPageIndex = calcCurrentPageIndex(scrollLeft, pagesScrollPositions);
 
@@ -542,6 +567,7 @@ const useSlideshowStyles = createUseStyles((theme) => ({
     },
     item: {
         width: '100%',
+        scrollSnapStop: isAndroid(theme.platformOverrides) ? 'always' : 'normal',
         scrollSnapAlign: 'start',
         flexShrink: 0,
     },
@@ -640,8 +666,10 @@ export const Slideshow: React.FC<SlideshowProps> = ({
         }
     }, [items.length]);
 
+    const shouldAutoplay = useShouldAutoplay(!!autoplay, carouselRef);
+
     React.useEffect(() => {
-        if (autoplay) {
+        if (shouldAutoplay && autoplay) {
             const time = typeof autoplay === 'boolean' ? DEFAULT_AUTOPLAY_TIME : autoplay.time;
             const loop = typeof autoplay === 'object' && autoplay.loop;
             const interval = setInterval(() => {
@@ -653,7 +681,7 @@ export const Slideshow: React.FC<SlideshowProps> = ({
             }, time);
             return () => clearInterval(interval);
         }
-    }, [autoplay, goNext, scrollRight]);
+    }, [autoplay, goNext, scrollRight, shouldAutoplay]);
 
     React.useEffect(() => {
         if (onPageChange) {