+++ src/video.tsx
@@ -2,6 +2,9 @@ import * as React from 'react';
 import {useDisableBorderRadius} from './image';
 import {createUseStyles} from './jss';
 import {combineRefs} from './utils/common';
+import {getPrefixedDataAttributes} from './utils/dom';
+
+import type {DataAttributes} from './utils/types';
 
 export type AspectRatio = '1:1' | '16:9' | '12:5' | '4:3';
 
@@ -50,6 +53,7 @@ export type VideoProps = {
     children?: void;
     /** defaults to none */
     preload?: 'none' | 'metadata' | 'auto';
+    dataAttributes?: DataAttributes;
 };
 
 const Video = React.forwardRef<HTMLVideoElement, VideoProps>(
@@ -62,6 +66,7 @@ const Video = React.forwardRef<HTMLVideoElement, VideoProps>(
             loop = true,
             preload = 'none',
             aspectRatio = '1:1',
+            dataAttributes,
             ...props
         },
         ref
@@ -118,6 +123,7 @@ const Video = React.forwardRef<HTMLVideoElement, VideoProps>(
                 preload={preload}
                 // This transparent pixel fallback avoids showing the ugly "play" image in android webviews
                 poster={poster || TRANSPARENT_PIXEL}
+                {...getPrefixedDataAttributes(dataAttributes)}
             >
                 {sources.map(({src, type}, index) => (
                     <source key={index} src={src} type={type} />